Add identifier/slug to strategies
